fix(notification): change employeeId type from int to Long in NotificationDto and NotificationRepo

+++ Tick-It-Notification-Service/src/main/java/com/harish/tickit/notificationService/repos/NotificationRepo.java
@@ -8,8 +8,8 @@ import com.harish.tickit.notificationService.models.Notification;
 @Repository
 public interface NotificationRepo extends JpaRepository<Notification, Integer>
 {
-	int countByEmployeeIdAndReadFalseAndDeletedFalse(int employeeid);
-	List<Notification> findByEmployeeIdAndReadFalseAndDeletedFalseOrderByCreatedAtDesc(int employeeid);
-	List<Notification> findByEmployeeIdAndDeletedFalseOrderByCreatedAtDesc(int employeeid);
+	int countByEmployeeIdAndReadFalseAndDeletedFalse(long employeeid);
+	List<Notification> findByEmployeeIdAndReadFalseAndDeletedFalseOrderByCreatedAtDesc(long employeeid);
+	List<Notification> findByEmployeeIdAndDeletedFalseOrderByCreatedAtDesc(long employeeid);
 
 }
